Change 'size' data type

+++ program/Flow.java
@@ -3,7 +3,7 @@ import java.util.LinkedList;
 public class Flow {
 
 	/* Attributes */
-	public double size;
+	public int size;
 	public int no_of_packets;
 	public int protocol;
 	public int start_time;
@@ -12,7 +12,7 @@ public class Flow {
 	public Flow() {
 	}
 
-	public Flow(double size, int no_of_packets, int protocol, int start_time){
+	public Flow(int size, int no_of_packets, int protocol, int start_time){
 		this.size = size;
 		this.no_of_packets = no_of_packets;
 		this.protocol = protocol;
@@ -22,7 +22,7 @@ public class Flow {
 	/* Methods */
 	public LinkedList<Packet> convertToPackets() {
 		LinkedList<Packet> packets = new LinkedList<Packet>();
-		int i, quotient = (int)size / no_of_packets;
+		int i, quotient = size / no_of_packets;
 		Packet p = new Packet(quotient, protocol, start_time);
 
 		for (i = 0; i < no_of_packets; i++) {
